only owner can delete channel from orbit db

+++ packages/backend/src/storage/storage.test.ts
@@ -158,7 +158,7 @@ describe('Storage', () => {
 })
 
 describe('Channels', () => {
-  it('deletes channel', async () => {
+  it('deletes channel as owner', async () => {
     storage = new Storage(tmpAppDataPath, 'communityId', { createPaths: false })
 
     const peerId = await createPeerId()
@@ -171,8 +171,32 @@ describe('Channels', () => {
 
     const eventSpy = jest.spyOn(storage, 'emit')
 
-    await storage.deleteChannel({ channelId: channelio.id })
+    await storage.deleteChannel({ channelId: channelio.id, ownerPeerId: peerId.toString() })
 
+    const channelFromKeyValueStore = storage.channels.get(channelio.id)
+    expect(channelFromKeyValueStore).toBeUndefined()
+    expect(eventSpy).toBeCalledWith('channelDeletionResponse', {
+      channelId: channelio.id
+    })
+  })
+
+  it('delete channel as standard user', async () => {
+    storage = new Storage(tmpAppDataPath, 'communityId', { createPaths: false })
+
+    const peerId = await createPeerId()
+    const libp2p = await createLibp2p(peerId)
+
+    await storage.init(libp2p, peerId)
+
+    await storage.initDatabases()
+    await storage.subscribeToChannel(channelio)
+
+    const eventSpy = jest.spyOn(storage, 'emit')
+
+    await storage.deleteChannel({ channelId: channelio.id, ownerPeerId: 'random peer id' })
+
+    const channelFromKeyValueStore = storage.channels.get(channelio.id)
+    expect(channelFromKeyValueStore).toEqual(channelio)
     expect(eventSpy).toBeCalledWith('channelDeletionResponse', {
       channelId: channelio.id
     })
